added custom scope 'SimpleThreadScope'
