+++ src/main/webapp/WEB-INF/applicationContext.xml
@@ -17,7 +17,7 @@
 
     <context:component-scan base-package="org.ayfaar.app"/>
 
-    <aop:aspectj-autoproxy/>
+    <aop:aspectj-autoproxy proxy-target-class="true"/>
 
     <import resource="classpath:spring-basic.xml"/>
     <import resource="classpath:hibernate.xml"/>
